Kannada Government Scheme Assistant/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.1
pymongo==4.6.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
